third_party/WebKit/LayoutTests/fast/dom/custom/lifecycle-created-parser-script.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script>
description("This test ensures that the lifecycle callback of a parser-made element is visible in following script block.")
proto = Object.create(HTMLElement.prototype, { createdCallback: { value: function() { window.callbacksCalled = true; } } });
document.registerElement("x-foo", { prototype: proto });
</script>
</head>
<body>
<x-foo></x-foo>
<script>
shouldBeTrue("window.callbacksCalled");
</script>
</body>
</html>
